requirements.txt
streamlit
dnspython
pandas
openpyxl
